First Nix configuration commit.
This is in an orbstack 'VM', so the boilerplate in the configuration.nix
might be strange as a result. In any case - this is just the beginning
of a larger goal to manage all of my configuration in a declarative way
with nix, allowing me to essentially package up my dotfiles/config and
load them onto any other machine effortlessly.