Nix updates: Moved Tmux config into nix!
I opted to just read my tmux.conf into the tmux.nix, so that my config is
exactly how I'm used to. Over time it would be better to migrate things into
the tmux flake configuration options directly.

In order to be able to read the tmux.conf in, I had to move it into a subdirectory
of the flake.nix, so I just `git mv`'d the .tmux.conf into home-manager/modules/tmux.conf
so it can be read into tmux.nix directly.

It seems to work. Hooray!