Formatting changes
Updated conform.nvim to prefer the LSP formatter if it is available.

This switches to lua_ls for formatting instead of stylua.

I may change this later, but for now, getting conform to pass command
line flags was problematic.

clang-format was doing the wrong thing when formatting google3 files.

Asking ciderlsp to do the formatting works though.

Also:
- Added PrettyPrint function that can print any lua table or type in
  notify.
- extended notify timeout to 20s.